♻️ modal, alert 창 컴포넌트 리팩토링
