+++ src/components/modal/Alert.scss
@@ -1,4 +1,4 @@
-@import "../../../components/style/default.scss";
+@import "../../components/style/default.scss";
 
 @keyframes showAlert {
   0% {
@@ -14,7 +14,7 @@
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  z-index: 100;
+  z-index: 999;
   justify-content: center;
   align-items: center;
   font-family: $medium-font;
