--- src/components/modal/alert/alertBase/AlertBtn.jsx
@@ -1,11 +1,0 @@
-function AlertBtn({ handleClick, children }) {
-  return (
-    <li>
-      <button type="button" onClick={handleClick} className="btn-do">
-        {children}
-      </button>
-    </li>
-  );
-}
-
-export default AlertBtn;
