--- src/components/modal/modals/PostModal.jsx
@@ -1,68 +1,0 @@
-import { useState, useContext } from "react";
-import UserContext from "../../../context/UserContext";
-import ModalLink from "../modalBase/ModalLink";
-import ModalBtn from "../modalBase/ModalBtn";
-import ModalFrame from "../modalBase/ModalFrame";
-import DeleteAlert from "../alert/alerts/DeleteAlert";
-import ReportAlert from "../alert/alerts/ReportAlert";
-
-//Post의 Moretn 클릭했을 떄 나타나는 모달
-function PostModal({ content, setOnModal, remove, from }) {
-  const { myAccountname } = useContext(UserContext);
-  const [onAlert, setOnAlert] = useState(false);
-  const accountname = content.author.accountname;
-  const postId = content.id;
-
-  function handleAlert() {
-    setOnAlert(!onAlert);
-  }
-
-  function handleCancel() {
-    setOnModal(false);
-    setOnAlert(false);
-  }
-
-  return (
-    <>
-      {myAccountname === accountname ? (
-        <>
-          {onAlert && (
-            <>
-              <DeleteAlert
-                text="게시글을 삭제하시겠어요?"
-                handleCancel={handleCancel}
-                // DeleteAlert 내 remove 함수에 연결되는 backUrl
-                backUrl={`post/${postId}`}
-                //삭제 후 리렌더링 위해 내려준 props
-                remove={remove}
-                from={from}
-              />
-            </>
-          )}
-          <ModalFrame setOnModal={setOnModal}>
-            <ModalBtn handleClick={handleAlert}>삭제</ModalBtn>
-            <ModalLink handleLink={`/upload/${postId}`}>수정</ModalLink>
-          </ModalFrame>
-        </>
-      ) : (
-        <>
-          {onAlert && (
-            <>
-              <ReportAlert
-                handleCancel={handleCancel}
-                // ReportAlert 내 report 함수에 연결되는 backUrl
-                backUrl={`post/${postId}/report`}
-                remove={remove}
-              />
-            </>
-          )}
-          <ModalFrame setOnModal={setOnModal}>
-            <ModalBtn handleClick={handleAlert}>신고</ModalBtn>
-          </ModalFrame>
-        </>
-      )}
-    </>
-  );
-}
-
-export default PostModal;
